I migrate confetti to success page.
