+++ components/GameScreen/GameScreen.tsx
@@ -7,31 +7,6 @@ import Image from "next/image"
 import PopupModal from "../PopupModal"
 import MintButton from "../MintButton"
 
-const client = new NFTStorage({
-  token: `${process?.env?.NEXT_PUBLIC_NFT_STORAGE_TOKEN}`,
-})
-const MUSIC_URLS = {
-  drums: [
-    "https://defient-music.s3.amazonaws.com/90_DRUMS_HIPHOP_001-SWEETS.wav",
-    "https://defient-music.s3.amazonaws.com/120_DRUMS_RNB_001-KATEE.wav",
-    "https://defient-music.s3.amazonaws.com/120_DRUMS_RNB_002-JOE.wav",
-  ],
-  vocal: [
-    "https://defient-music.s3.amazonaws.com/120_MUSIC_Bbmin_RNB_002-MCKENNA.wav",
-    "https://defient-music.s3.amazonaws.com/120_MISC_RNB_001-JEFF.wav",
-    "https://defient-music.s3.amazonaws.com/120_MUSIC_Bbmin_RNB_001-SERENA.wav",
-  ],
-  bass: [
-    "https://defient-music.s3.amazonaws.com/120_MUSIC_Bbmin_RNB_002-MCKENNA.wav",
-    "https://defient-music.s3.amazonaws.com/90_DRUMS_HIPHOP_002-SAMEER.wav",
-    "https://defient-music.s3.amazonaws.com/90_DRUMS_TRAP_001-MARION.wav",
-  ],
-  guitar: [
-    "https://defient-music.s3.amazonaws.com/90_MISC_HIPHOP_001-SHANE.wav",
-    "https://defient-music.s3.amazonaws.com/90_MUSIC_Bbmin_HIPHOP_001-MICHAEL.wav",
-    "https://defient-music.s3.amazonaws.com/90_MUSIC_Bbmin_HIPHOP_002-DOM.wav",
-  ],
-}
 interface IOption {
   id: string
   name: string
@@ -51,8 +26,6 @@ const GameScreen = ({ onSuccess }: any) => {
     bass: false,
     guitar: false,
   })
-  const [cid, setCid] = useState<string>("")
-  const [mixing, setMixing] = useState<boolean>(false)
   const options: IOption[] = [
     { id: "bass", name: "Bass", imgUrl: "/bass.png" },
     { id: "drums", name: "Drums", imgUrl: "/drums.png" },
@@ -67,26 +40,6 @@ const GameScreen = ({ onSuccess }: any) => {
     }
     setChecked({ ...checked, [value]: !checked[value] })
   }
-  const remixHandler = async () => {
-    setMixing(true)
-    const crunker = new Crunker()
-    const buffers = await crunker.fetchAudio(
-      _.sample(MUSIC_URLS[choices[0]]),
-      _.sample(MUSIC_URLS[choices[1]]),
-    )
-    const merged = await crunker.mergeAudio(buffers)
-    const output = await crunker.export(merged, "audio/wav")
-    const CID = await client.storeBlob(output.blob)
-    setCid(CID)
-    setMixing(false)
-    setChoices([])
-    setChecked({ drums: false, vocal: false, bass: false, guitar: false })
-  }
-  useEffect(() => {
-    if (cid) {
-      toast.success(`🎉 Remix created!, check it out at ipfs://${cid}`)
-    }
-  }, [cid])
 
   return (
     <div className="flex flex-col items-center justify-center h-screen gap-4 align-center">
@@ -105,19 +58,8 @@ const GameScreen = ({ onSuccess }: any) => {
           </button>
         ))}
       </div>
-      {choices.length > 1 && (
-        <div>
-          <button
-            type="button"
-            className="p-4 m-2 border-4 border-purple-600 rounded-full"
-            onClick={remixHandler}
-          >
-            Remix
-          </button>
-        </div>
-      )}
-      {cid && <MintButton cid={cid} onSuccess={onSuccess} />}
-      {mixing && <PopupModal open={mixing} />}
+
+      {choices.length > 1 && <MintButton choices={choices} onSuccess={onSuccess} />}
     </div>
   )
 }
